pyproject.toml
[project]
name = "pyparquetdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=1.1.3",
    "filelock>=3.16.1",
    "flask>=3.1.0",
    "pandas>=2.2.3",
    "polars>=1.13.1",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.0.1",
]
